Created IsUnmodifiableCollection class to improve mismatch description
The mismatch description provided by anyOf() is not helpful (it just
calls toString() on the item being tested).  This is not a fault but
just a limitation of what it can do in the general case.

By providing a custom IsUnmodifiableCollection Matcher class we can
produce a better mismatch description.